DFI-1820 Scala3 Upgrade
* DFI-1820 Scala3 Upgrade

+++ app/uk/gov/hmrc/disareturns/repositories/NotificationContextRepository.scala
@@ -41,7 +41,7 @@ class NotificationContextRepository @Inject() (mc: MongoComponent, appConfig: Ap
           keys = Indexes.ascending("updatedAt"),
           indexOptions = IndexOptions()
             .name("updatedAtTtlIdx")
-            .expireAfter(appConfig.timeToLive, TimeUnit.DAYS)
+            .expireAfter(appConfig.timeToLive.toLong, TimeUnit.DAYS)
         )
       ),
       replaceIndexes = true
